pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_controller"
version = "0.1.0"
readme = "README.md"
dependencies = ["python-dotenv", "langchain", "langchain-openai", "langgraph", "trustcall", "IPython", "langchain-chroma"]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests", "tests.*", "docs", "build", "dist", "config"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.txt"]

[project.scripts]